Remove the stubPerforceService and update all tests to use the StubPerforceModel
